fix: 자료형 정리

+++ maeil-wiki/src/main/java/maeilwiki/mutiplechoice/application/MultipleChoiceWorkBookRequest.java
@@ -4,10 +4,10 @@ import java.util.List;
 
 public record MultipleChoiceWorkBookRequest(
         String workBookTitle,
-        Long difficultyLevel,
+        int difficultyLevel,
         String category,
         String workbookDetail,
-        Long timeLimit,
+        Integer timeLimit,
         List<MultipleChoiceQuestionRequest> questions
 ) {
 }
